% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsOrder.R
\name{acsOrder}
\alias{acsOrder}
\title{Order the observations of an acs object}
\usage{
acsOrder(acs, FUN = acs::standard.error, col = 1)
}
\arguments{
\item{acs}{An \code{acs} object}

\item{FUN}{A function for extracting information from an \code{acs} object (e.g. \code{estimate}, \code{standard.error})}

\item{order_col}{The column to order by}

\item{result_col}{The column to display, as a numeric index or as a character string}
}
\description{
A convenience function for ordering acs objects
}

