% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-geom.R
\name{coerce_geom}
\alias{coerce_from_geometrycollection}
\alias{coerce_geom}
\alias{coerce_to_geom}
\title{A set of functions for working with sf objects}
\usage{
coerce_from_geometrycollection(x, crs = NULL)

coerce_to_geom(x, FUN, crs = NULL)
}
\arguments{
\item{x}{A sf-class data object}

\item{crs}{A crs-class object (coordinate reference system)}

\item{FUN}{A function for creating a sfg-class object (e.g., sf::multipolygon)}
}
\description{
Functions for working with sf objects.
}

