% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\docType{data}
\name{crs}
\alias{crs}
\alias{crs_proj}
\alias{crs_wspn}
\title{Coordinate reference systems}
\format{An object of class \code{CRS} of length 1.}
\usage{
crs_proj

crs_wspn
}
\description{
A set of coordinate reference systems.
}
\keyword{datasets}
