% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_convert.R
\name{currency_convert}
\alias{currency_convert}
\title{An currency conversion function}
\usage{
currency_convert(x, target_year, base_year)
}
\arguments{
\item{x}{A dollar amount}

\item{target_year}{The year to convert the dollar value to}

\item{base_year}{The year to convert the dollar value to}
}
\description{
A one-to-one spatial overlay function, similar to sp::over()
}
