% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_convert_rate.R
\name{currency_convert_rate}
\alias{currency_convert_rate}
\title{An currency conversion function}
\usage{
currency_convert_rate(target_year, base_year)
}
\arguments{
\item{target_year}{The year to convert the dollar value to}

\item{base_year}{The year to convert the dollar value to}
}
\description{
A one-to-one spatial overlay function, similar to sp::over()
}
