% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_read.R
\name{drive_read}
\alias{drive_read}
\title{Read a file from Google Drive.}
\usage{
drive_read(dr_id, read_fun, .unzip = FALSE, folder_name = NULL, ...)
}
\arguments{
\item{dr_id}{a Google Drive object id (use \code{\link[googledrive]{as_id}}).}

\item{read_fun}{a function to read the object into memory.}

\item{.unzip}{A logical scalar, does the file need to be decompressed?}

\item{folder_name}{If the target is \code{.zip} file, what is the name of the folder within the file.}

\item{...}{Arguments to pass to the `read_fun` function.}
}
\value{
an object, format depends on which `read_fun` is provided.
}
\description{
Read a file from Google Drive.
}
