% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload_obj.R
\name{drive_upload_obj}
\alias{drive_upload_obj}
\title{Upload a object in the Working Environment to Google Drive.}
\usage{
drive_upload_obj(x, filename, write_fun, .zip = FALSE, .share = TRUE)
}
\arguments{
\item{x}{an object in the working environment.}

\item{filename}{Character, name of the file (including appropriate extension) that
should appear on Google Drive.}

\item{write_fun}{a function to write the object into a temporary file}

\item{.zip}{A logical scalar, do you want to compress the file into a `.zip` file?}

\item{.share}{A logical scalar, do you want to share the file?}
}
\value{
`dribble` object
}
\description{
Upload a object in the Working Environment to Google Drive.
}
