% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMakeShpObjectTemplate.R
\name{insertMakeShpObjectAddin}
\alias{insertMakeShpObjectAddin}
\title{An RStudio add-in for inserting spatial data loading code}
\usage{
insertMakeShpObjectAddin()
}
\description{
An add-in that inserts the code to build and/or load a spatial object in R.
}

