% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_id.R
\name{make_id}
\alias{make_id}
\title{Make a unique character string for anonymous record identification}
\usage{
make_id(rn, string = LETTERS, .shuffle = FALSE)
}
\arguments{
\item{rn}{number of records (e.g. \code{nrow(df)})}

\item{string}{a vector of characters to form the identifier string}

\item{.shuffle}{logical, whether to shuffle the ids or keep them in order}
}
\description{
Make a unique character string for anonymous record identification
}
