% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_or_read.R
\name{make_or_read}
\alias{make_or_read}
\title{Read or create a file}
\usage{
make_or_read(fp, make_expr, read_expr)
}
\arguments{
\item{fp}{filepath}

\item{make_expr}{expression used to create the file}

\item{read_expr}{expression used to read the file}
}
\value{
The object that is read or created is returned silently.
}
\description{
A function that determines whether to read or create a file based on whether that file currently exists.
}
