% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscgis_pals.R
\docType{data}
\name{miscgis_pals}
\alias{miscgis_pals}
\title{Mapping Color Palettes}
\format{An object of class \code{list} of length 1.}
\usage{
miscgis_pals
}
\description{
A set of color palettes.
}
\keyword{datasets}

