% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myAcsPlot.R
\name{myAcsPlot}
\alias{myAcsPlot}
\title{Ordered Dotplot with Error Bars}
\usage{
myAcsPlot(acs, sort_fun, col = 1, ylab, .pct = TRUE)
}
\arguments{
\item{acs}{An \code{acs} object}

\item{sort_fun}{A function for extracting information from an \code{acs} object (e.g. \code{estimate}, \code{standard.error}), which will be used to sort the \code{acs} object.}

\item{col}{The column to display, as a numeric index or as a character string}

\item{ylab}{= A label for the y-axis (should indicated the units, as in 'People of Color (%)')}

\item{.pct}{= T/F, whether the units are a percentage or not}
}
\description{
An ordered dotplot with error bars and labels
}
