% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLfltOpts.R
\name{myLfltOpts}
\alias{myLfltOpts}
\title{Options for styling a leaflet map}
\usage{
myLfltOpts(map, bumpLabels = TRUE, hideControls = TRUE)
}
\arguments{
\item{bumpLabels}{Include OSM labels (styled by CartoDB) above all other layers (including paths).}

\item{hideControls}{Show leaflet control elements (e.g., zoom buttons, attribution, etc.)}
}
\description{
A set of style options to be applied to a leaflet htmlwidget using a JS script.
}

