% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySptlPolyDF.R
\name{mySptlPolyDF}
\alias{mySptlPolyDF}
\title{Convert a SpatialPolygons object into a SpatialPolygonsDataFrame}
\usage{
mySptlPolyDF(sp)
}
\description{
\code{mySptlPolyDF} converts a SpatialPolygons object into a SpatialPolygonsDataFrames.
}
