% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_sf.R
\name{over_sf}
\alias{over_sf}
\title{An sf implementation of sp::over()}
\usage{
over_sf(x, y)
}
\arguments{
\item{x, y}{An sf-class object}
}
\description{
A one-to-one spatial overlay function, similar to sp::over()
}
