% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lookup_string.R
\name{parse_lookup_string}
\alias{parse_lookup_string}
\title{Convert a structured character string into a named vector.}
\usage{
parse_lookup_string(string, col_sep, row_sep)
}
\arguments{
\item{string}{A character string containing codes and their corresponding values.}

\item{col_sep}{A character string indicating a new column (e.g, \code{"\\s=\\s"}).}

\item{row_sep}{A character string indicating a new row (e.g, \code{"\n"}).}
}
\description{
A function for converting a character string into a named vector.
}
