% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_round.R
\name{smart_round}
\alias{smart_round}
\title{Round a vector of numerics while preserving their sum}
\usage{
smart_round(x, digits = 0)
}
\arguments{
\item{x}{a vector of numeric values}

\item{digits}{the number of digits to round each value to}
}
\description{
Round a vector of numerics while preserving their sum
}
\note{
This function comes from a \href{https://stackoverflow.com/a/35930285/5170276}{post on Stack Overflow}.
}
