% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_drop_geometry.R
\name{st_drop_geometry}
\alias{st_drop_geometry}
\title{Read a file from Google Drive.}
\usage{
st_drop_geometry(x)
}
\arguments{
\item{x}{an \code{sf} object}
}
\value{
a \code{tibble} object without the \code{geometry} column
}
\description{
Read a file from Google Drive.
}
