% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_intersects_any.R
\name{st_intersects_any}
\alias{st_intersects_any}
\title{Do any spatial intersections exist?}
\usage{
st_intersects_any(x, y)
}
\arguments{
\item{x, y}{\code{sf} objects}
}
\value{
logical
}
\description{
A function that determines whether any spatial intersections exist between two \code{sf} objects.
}
