% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_nest_sf.R
\name{st_nest_sf}
\alias{st_nest_sf}
\title{Create an sf object with metadata contained in a nested tibble}
\usage{
st_nest_sf(x)
}
\arguments{
\item{x}{an \code{sf} object}
}
\value{
an \code{sf} object with metadata columns contained in a nested tibble (by row).
}
\description{
A function that returns an sf object with metadata columns contained in a nested tibble (by row).
}
\examples{

library(sf)
nc <- st_read(system.file("gpkg/nc.gpkg", package = "sf"), quiet = TRUE)

\dontrun{
  st_nest_sf(nc)
}
}
