% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleWidget.R
\name{styleWidget}
\alias{styleWidget}
\title{Prepend CSS Styling to a HTML Widget}
\usage{
styleWidget(hw = NULL, style = "", addl_selector = "")
}
\arguments{
\item{hw}{An HTML widget}

\item{style}{A string of CSS style choices}

\item{addl_selector}{An additional selector to target a specific element}
}
\description{
Prepends a string of CSS styles to an HTML widget
}
\details{
Source: http://widgets1092.rssing.com/browser.php?indx=61900561&last=1&item=19
}
