% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{cbind_fill}
\alias{first_not_na}
\alias{is_pct}
\alias{label_pct}
\alias{label_yr}
\alias{mk_proj_dir}
\alias{replace_empty}
\alias{subset_duplicated}
\alias{utils}
\title{A set of generic helpful functions}
\usage{
first_not_na(x)

replace_empty(x)

subset_duplicated(x, nm, notin = FALSE)

cbind_fill(...)

mk_proj_dir()

is_pct(x)

label_pct(breaks)

label_yr(breaks)
}
\description{
General purpose functions to make working in R easier.
}

