% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_read.R
\name{drive_read}
\alias{drive_read}
\title{Read a file from Google Drive.}
\usage{
drive_read(dribble, read_fun, .unzip = FALSE)
}
\arguments{
\item{dribble}{a `dribble` object containing the googledrive information.}

\item{read_fun}{a function to read the object into memory.}

\item{.unzip}{A logical scalar, does the file need to be decompressed?}
}
\value{
an object, format depends on which `read_fun` is provided.
}
\description{
Read a file from Google Drive.
}
