% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLflt.R
\name{myLflt}
\alias{myLflt}
\title{A pleasant leaflet basemap}
\usage{
myLflt(tile_opts = tileOptions(), chinatown = FALSE)
}
\arguments{
\item{tile_opts}{list, override arguments to the defaults in \code{\link[tileOptions]{leaflet}}}

\item{chinatown}{logical, whether to use Hing Hay Park as the default \code{\link[setView]{leaflet}}}
}
\description{
A pleasant-looking leaflet basemap to begin a leaflet pipechain with.
}
\details{
Note: because this uses tiles from a Mapbox URL, this function will not work correctly
in a PDF document - use \code{myLflyGrey} instead.
}
