% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLfltGrey.R
\name{myLfltGrey}
\alias{myLfltGrey}
\title{A subtle grayscale leaflet basemap}
\usage{
myLfltGrey(bumpLabels = TRUE, hideControls = TRUE)
}
\arguments{
\item{bumpLabels}{Include OSM labels (styled by CartoDB) above all other layers (including paths).}

\item{hideControls}{Show leaflet control elements (e.g., zoom buttons, attribution, etc.)}
}
\description{
A subtle grayscale leaflet basemap to begin a leaflet pipechain with.
}

