% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLfltOpts.R
\name{myLfltOpts}
\alias{myLfltOpts}
\title{Options for styling a leaflet map}
\usage{
myLfltOpts(map, tileLabels = TRUE, fullScreenBtn = TRUE,
  bumpTileLabels = TRUE, hideControls = TRUE)
}
\arguments{
\item{map}{The map to add the tile layer to.}

\item{tileLabels}{If \code{tileLabels} is \code{TRUE}, a tile layer of OSM labels (styled by CartoDB) will be added.}

\item{fullScreenBtn}{If \code{fullScreenBtn} is \code{TRUE}, a fullscreen control button will be added.}

\item{bumpTileLabels}{If \code{bumpTileLabels} is \code{TRUE}, the OSM labels will sit above all other map layers.}

\item{hideControls}{If \code{hideControls} is \code{TRUE}, control elements will be hidden.}
}
\description{
A set of frequently-used leaflet map options.
}
